Add more information to the WTP protocol summary (PDU type and PDU specific important parameters).
Document the computation of the length field in WTP concatenation after having
looked at a capture with the length field encoded as a WSP uintvar-integer.

Use "common code" for WTP reassembly, by calling process_reassembled_data().
Document the behavior of reassembly as the output of Ethereal differs between
the first and the second pass.

Question: shoud the common reassembly code provide a call-back mechanism to
get access to previously-unreassembled packets that appear to be part of a
reassembled whole, and to be able to update the state and information of
those packets at the time of the reassembly?


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@9547 f5534014-38df-0310-8fa8-9805f1628bb7