Add a routine to be used to put up alert boxes for invalid display filter expressions; use that in a number of places, so we use the same alert box.  (More work is needed to figure out the right way to handle some other "dfilter_compile()" failures.)
Use the error message from the display filter as the primary error, as
that's the message that tells you what the underlying problem is.  (The
GNOME HIG says "In most situations the user should only need the primary
text to make a quick decision", so the primary text should tell you
what's wrong with the filter, not just that it's invalid.  If there are
messages from the display filter code that don't give enough
information, or are a bit cryptic, such as "Unexpected end of filter
string," those should be fixed in the display filter code.)

Improve the error used if an empty filter is used for "find frame".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@10025 f5534014-38df-0310-8fa8-9805f1628bb7