* Color filters:
Read of color filters rewritten to remove possibility of buffer overflow
in read_filters_file() and to fix comment & blank line handling.
This removes any limitation (and stack overflow) on the lengths of color
filter name and filter expression which are now dynamically allocated.

Fix removal/deletion of color filters (avoid segmentation fault when
some color filter is deleted and then the list cleared for instance).
The use of g_slist_remove() within an iterator is simply unsafe if the
code is not prepared to handle this.

Fix a small memory leak in revert_filters().

Remove inappropriate comments in read_filters().

Miscellaneous code cleaning.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@10380 f5534014-38df-0310-8fa8-9805f1628bb7