Rename a bunch of variables and routines that pertain to string search with "string" rather than "ascii", to make it clearer what they're involved with.
Use "gtk_toggle_button_set_active()", not
"gtk_toggle_button_set_state()" (the latter is a deprecated alias for
the former, probably dating back to GTK+ 1.0[.x] - 1.2[.x] and later
have "gtk_toggle_button_set_active()").

Do *NOT* change the radio buttons for the type of string search to do
based on whether we're doing a string search or not - doing so means we
don't correctly remember the type of string search.

Get rid of code to fetch some values that we don't subsequently use.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@10609 f5534014-38df-0310-8fa8-9805f1628bb7