Add a discriminator for WV-CSP so the correct version is used for rendering the WBXML tokens in WV-CSP.
Provide rendering of OPAQUE data for media types, based on a framework of 4 functions that take care of well-known and literal tags and attribute names for deciding on the meaning of the OPAQUE data. Applied to SI, EMN and WV-CSP.

Update reference URLs for WAP, OMA and related specs.

Note: as the WBXML common code only takes ~2000 lines while the token mappings take ~4000 lines, maybe it is time to envisage writing the per-media mappings to dedicated source files (packet-wbxml-wml.c, packet-wbxml-wv_csp.c etc)?


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@10615 f5534014-38df-0310-8fa8-9805f1628bb7