Use "tvb_reported_length_remaining()" in "until the end of the packet" loops, so we mark frames as short if they've been cut short by a snapshot length.
The user class in buddy list TLVs appears to be 2 bytes, in at least
some captures; make "dissect_aim_userclass()" take the user class value,
and the length of the user class field, as arguments, and have the
caller fetch the value.  Also, display the numerical value of the user
class in hex, as it's a bitset.

Fix the messaging dissector to put stuff under the top-level item rather
than at the top level.

Fix a typo.

Clean up indentation.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@10802 f5534014-38df-0310-8fa8-9805f1628bb7