Make "-q" suppress packet printing when you're reading a capture file and not writing to another capture file, for use with "-z" options.
Note that "-z proto" *doesn't* print statistics at the end - it modifies
the packet summary output.

Note that on at least some BSDs the "status" character is set to "off"
by default, so you have to set it explicitly in order to be able to ^T
an application such as Tethereal.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11038 f5534014-38df-0310-8fa8-9805f1628bb7