Give the RFC number for ARP.
When checking for a gratuitous ARP, compare the raw bytes of the
protocol addresses, not the strings for those addresses.

Do the stuff we do even if we *aren't* constructing a protocol tree or
setting the Info column first, and then quit if we're doing neither.
That obviates the need to set "is_gratuitous" if we're doing neither.

Construct the strings for addresses when we need them, rather than
constructing them in advance even if we don't need them.

Capitalize "ARP" in "Gratuitous ARP".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11168 f5534014-38df-0310-8fa8-9805f1628bb7