From Martin van der Werff: don't do "tvb_ensure_length_remaining()" to create the parameter tvbuff, as that requires that there's at least one byte of parameter data, and some messages have no parameters and would cause an exception to be thrown in that case.
Just use -1 for the length parameters when creating the parameter
tvbuff, so that it includes the full reported length.

Don't put undissected parameter information into the protocol tree if
there aren't any parameters - and don't bother getting the number of
bytes of parameter information unless we're putting undissected
parameter information into the protocol tree and need it to see whether
there are any parameters.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11297 f5534014-38df-0310-8fa8-9805f1628bb7