In "packet-ieee80211.c", use the same scheme for displaying the data rate used elsewhere, rather than using floating point.
In "packet-prism.c", show the data rate correctly in the column display,
and show it as Mb/s in the protocol tree, and show it as "Data Rate", as
it is in "packet-ieee80211.c".

In "packet-radiotap.c" and "packet-wlancap.c", show it as "Data Rate"
rather than as "Datarate".

Show the units uniformly as "Mb/s", rather than "mb/s" (it's mega, not
milli) or "Mbps", except for the AVS header, where it's Kb/s.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11323 f5534014-38df-0310-8fa8-9805f1628bb7