Don't allow non-positive values for the main window width and height values in the "recent" file.  Also, check for syntax errors in numerical values.
Set the main window height, as well as the main window width, to default
values before we read in the values from the "recent" file.  (Use
non-positive values here if you want the default to be "don't set the
geometry" - note that you can't set one to a positive value and the
other one to a non-positive value and expect only one to be set; they
*both* have to be positive in order for the size to be set.)

Treat non-positive values for those settings in the "recent" structure
as an indication that we don't have values from them and that we should
therefore not set the window size.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11353 f5534014-38df-0310-8fa8-9805f1628bb7