fix to that long outstanding bug with windowscaling
If window scaling is NOT offered in the SYN+ACK  then window scaling will
not be used at all, so clear it if we saw it offered previously in the SYN packet.

If the window is scaled in a packet,  make ethereal display that by appendign the 
string "  (scaled)" to the end of the tcp.window line in the
decode pane.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11837 f5534014-38df-0310-8fa8-9805f1628bb7