From Martin Mathieson:  I've written this patch to use the 'Delay since last SR' (DLSR) field found in SR reports to calculate and report roundtrip-propagation delays.  This is described in rfc 3550, section 6.4.1, inside the description of DLSR.
Only the endpoint can compute the end-end roundtrip delay, and only they
know exactly when the report is received and can compare it with the 'Last
SR timestamp' (LSR) that they set.  This patch instead takes the difference
between the capture times of the 2 reports and subtracts the DLSR (the LSR
is checked in case the SR it's referring to wasn't captured).  The time
difference represents a roundtrip network delay between the point of capture
and the sender of the SR containing the DLSR.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@11998 f5534014-38df-0310-8fa8-9805f1628bb7