from Mark Phillips: the length field in the UDP header includes the length of the UDP header itself, so subtract the length of the header when using it to limit the length of the payload tvbuff.
Clean up the computing of the captured length of the payload tvbuff (we
really should get rid of the "length" argument to "tvb_new_subset()",
and have it compute the captured length based on the supplied reported
length and the amount of that data actually present in the parent
tvbuff).

Don't fetch the length and checksum fields until we use them (so that we
don't throw an exception until then, and fail to process the source and
destination ports), and check whether the length is bogus regardless of
whether we're building a protocol tree or not.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@12444 f5534014-38df-0310-8fa8-9805f1628bb7