If an item in a source description chunk begins with 4 bytes of zero, that doesn't mean it's padding at the end of a previous item - it might, for example, be the *first* item in the chunk.  Don't treat it as padding.
Do, however, treat an item that begins with a zero byte as an item, but
break out of the loop processing items as soon as the item type is put
into the protocol tree, as there's no length field or data in an
RTCP_SDES_END item.  Fix the comment for that loop to indicate that the
loop checks both for end-of-frame and for an RTCP_SDES_END item.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@13040 f5534014-38df-0310-8fa8-9805f1628bb7