From Jon Ringle:
	don't dissect attributes if there aren't any;

	put each attribute into a subtree;

	register the dissector by name.

Use "match_strval()" to check whether the message type is a known STUN
message type, and to generate the message type for the Info column.

Don't use "tvb_bytes_exist()" to check when we run out of data - use the
length fields from the packet.  Check the sanity of those lengths, too.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@13063 f5534014-38df-0310-8fa8-9805f1628bb7