Add the I-D name for L2TPv3 to a comment.
Do more, and more careful, AVP length checks.

Use tvb_format_text() to display strings.

Just use "col_add_fstr()" to set the Info column - there's no need to
format into a buffer and then set the column to that buffer.

Make the L2TP dissector a "new-style" dissector and have it reject
packets with an unknown version number - the dissection effect is the
same if no other dissector claims the packet (it just gets displayed as
data), but this allows other dissectors to claim it.

Don't use the length field if the length bit isn't set.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@13201 f5534014-38df-0310-8fa8-9805f1628bb7