Don't assume that the only blank filter expression is empty - check for an all-space-character string.
Don't use "filter_string" both for the (unfreeable, unmodifiable) value
fetched from the text entry and the new filter expression; use it only
for the former, and make it a "const" pointer to avoid compiler
warnings.

Use a GString to construct the stuff to be added to the new filter
expression - the old code, when appending stuff to the filter expression
in a loop, would leak the pre-append value.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@13237 f5534014-38df-0310-8fa8-9805f1628bb7