a lot of people dont specify top level pointers in teh idl and just cerlare it as a parameter that is a struct/union and not a ref pointer to one.
this is ok since toplevel ref pointers are invisible in the wire encoding anyway.

unfortunatelky  ethereal dce runtime needed to see that pointer to keep track of whisch one was a toplevel and which one wasnt.


implement a new api to call pointers which explicitely specifies whether the pointer is toplevel or embedded.
eventually all dce dissectors will use this new interface and the generic dissect_ndr_pointer() can be retired.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@13755 f5534014-38df-0310-8fa8-9805f1628bb7