Get rid of boilerplate.
Use "tvb_ensure_bytes_exist()" to check for some invalid lengths.

Add some comments about possible problems in the code.

Get rid of an unnecessary length check (the length is the sum of a small
constant and a value extracted from a single byte, so the sum won't
overflow).

For a FCP_RSP, make the top-level protocol tree item run to the end of
the tvbuff and then set its length when we finish dissecting it (if we
throw an exception and don't get around to setting the length, that
means that we hit the end of the tvbuff before we hit the end of the
item).  Add some checks to catch too-large length fields.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@13917 f5534014-38df-0310-8fa8-9805f1628bb7