Major speed improvement for filtering and dissection.
It should not dump core as far as all my tests are concerned and Menu_Statistics/ProtocolHierStats work

It needs more testing and there might still be cases where it will crash that will need to be fixed  but I feel it will be worth it since it will decrease the time to filter very large capture files dramatically.

Real significant performance boost for very large captures.


(If we cant fix all the problems we can just revert this patch)




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@14051 f5534014-38df-0310-8fa8-9805f1628bb7