With the speedup changes to epan/proto.c  the tree will be faked to only contain those fields actually referenced by a filter. Other nodes will never be created by the tree (sort of auto pruning of the tree).
While this greatly spped up the processing performance of ethereal, it makes it "dangerous" for apps that try to walk the tree directly.


rtp_analysis did not specify the fields "rtp.version" nor " rtp.ssrc" when dissecting the current packet  and as such these fields were no longer part of the tree.
This surprised rtp_analysis and a coredump resulted.


This change will add the field that we walkt the tree to find to the filter string so that they will be there.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@14200 f5534014-38df-0310-8fa8-9805f1628bb7