From Mike Duigou:  a patch which adds support for displaying jxta conversations and endpoints from the 'Statistics' menu. Also adds :
  - a generated field to Welcome messages to indicate whether this 
welcome is from the initiator or the receiver. You can filter on this 
field with 'jxta.welcome.initiator==[0|1]'

  - Marks the source and destination fields of message added in the last 
patch as generated fields. (Saw it being done for some other protocol).

  - Cleans up dissection of Message which used to assume it would 
dissected more than one set of hdr/message per tvbuff. Now it only 
attempts to dissect one.

- Uses GMemChunk for tap info and for conversation data rather than 
g_malloc() There's still a major leakage of g_malloc()ed the c-strings 
which are used in jxta addresses. Any suggestions for how these can be 
allocated such that they can be freed re-init is called would be 
appreciated. For address objects whose data ptr doesn't point into a 
tvbuff it's not clear what the lifecyle of an address object is.

  - Fixes a bug with filling in the of the transport layer conversation 
data. >= vs. >



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@14558 f5534014-38df-0310-8fa8-9805f1628bb7