Find the BitTorrent dissector immediately after registering it, and save the handle, rather than finding it every time we need it.
Just make a guint8 array with the BitTorrent magic string length
followed by the string, and use "tvb_memeql()" to test against that - it
handles checking to make sure there's enough data in the tvbuff (if
not, it returns -1 as a "no match" indication) and then checks all of
the first 20 bytes in one operation.

Clean up white space.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@14882 f5534014-38df-0310-8fa8-9805f1628bb7