packet_ber has for some time been doing sanity checking of the BER Length fields that they are not longer than the reported length of the tvb.
this triggers some bugs since in packet-ber we are a bit too lax in setting reported_length of the tvb_new_subset() tvb.

this cause short kerberos packets to not be decoded at all and the same for other short asn based packets as well.


fix some of these instances.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@15127 f5534014-38df-0310-8fa8-9805f1628bb7