Pick up some fixes and enhancements from http://cvs.fedora.redhat.com/viewcvs/rpms/ethereal/FC-4/:
In the LPD dissector, make lpd_client_code a value_string so that we
don't segfault.  Do the same for lpd_server_code, although it's not
strictly necessary.

Check to see if htmlview is installed, and use it as our HTML viewer.

The Fedora RPM has other patches, but I'm not sure if they should
be applied.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@15143 f5534014-38df-0310-8fa8-9805f1628bb7