Don't assign "" to a string that's only used after it's assigned to.
"tvb_fake_unicode()" never returns a null pointer, so there's no need to
check for it; there's also no need to check that it's not an empty
string before freeing it, as it's always assigned a g_malloc()ed string.

It also takes as an argument the number of Unicode characters in the
string, not the number of bytes in the string.

Use "format_text()" when adding a string to the Info column, in case
it has non-printable characters.

Don't use "proto_tree_add_string_format()" if you don't have to.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@15239 f5534014-38df-0310-8fa8-9805f1628bb7