From Mike Duigou:
A patch to the JXTA dissector. It's purpose is to improve 
the behaviour of the dissector when dissecting captures that do not 
include the whole conversation.

The dissector is now more capable of isolating a conversation without 
seeing the beginning of the conversation. It's still not always able to 
pick up all conversations (suggestions, patches and hints welcome).

improved some of the doxygen docs and comments.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@15500 f5534014-38df-0310-8fa8-9805f1628bb7