Use "format_text()" and "tvb_format_text()" when displaying strings. (It's not necessary, or desirable, to use them when fetching strings.)
When processing a VLAN TLV, check the length of the TLV, so we don't end
up running past the end of the TLV and getting a negative remaining
length.  Append to the protocol tree item for the TLV, and the Info
column, as we dissect the VLAN ID and VLAN name fields.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@15515 f5534014-38df-0310-8fa8-9805f1628bb7