remove all strcpy() from packet-fc.c
rewrite the functions to do proto_item_append_text() instead of building a string and then printing it.

The new function is functionally equivalent to the previous function except it does not print the values of the multi-bit fields to the expansion
(the expansion line is already a km wide as it is)



there are now only 202 strcpy() left in epan/dissectors  down from 300+ instances some weeks ago.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@16009 f5534014-38df-0310-8fa8-9805f1628bb7