Fix indentation.
Gate X and Y Offset are 16-bit integral numbers of degrees, not 32-bit
IEEE floating-point numbers, and the numbers in collision detection
segment definitions and collision detection volume definitions are
16-bit fixed-point numbers, not 32-bit IEEE floating-point numbers;
handle them as such - this fixes bug 605.

When displaying the packet length in the Info column, display the *real*
length, not the sliced length.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@16714 f5534014-38df-0310-8fa8-9805f1628bb7