From Grame Lunt:  updated X.500 dissectors to include DOP support.
The "dop" dissector is the renamed "x501" dissector consequently the asn/x501 directory should be removed. The patch includes the changes to epan/dissectors/Makefile.common to reflect this.

As the DOP dissection is not fully tested, I have disabled it by default for now (like DSP) but it can be enabled by the user.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@16726 f5534014-38df-0310-8fa8-9805f1628bb7