fix #648:
Win32 only: reading a pathname from an environment var requires us to read it in as unicode somehow and convert it to utf8. Using _wgetenv should work under all circumstances on NT, using getenv and g_locale_to_utf8 on Windows OT is the best we can do in this case.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17024 f5534014-38df-0310-8fa8-9805f1628bb7