From Greg Morris: epan/dissectors/ncp2222.py - Fixes the NCP group values for all NCP's. Also fixes some additional return values and cleanup.
gtk/ncp_stat.c - Fixes the NCP group values for SRT.

 gtk/service_response_time_table.c:

The SRT is broken if you hit the reload button or apply a filter. The table isn't cleared so each item in the list is duplicated and the second entries remain with initial values. This patch clears the GTK_CLIST so that the redundant entries no longer appear.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17139 f5534014-38df-0310-8fa8-9805f1628bb7