Add a routine to attempt to get the absolute pathname of the executable file, strip off the last component to get the pathname of the directory containing the executable file, and save it for future use.  On Windows, you can get that from the OS, but, on UN*X, you have to look at argv[0] and derive the absolute path from that (argv[0] is not guaranteed to be an absolute path, or even a path at all).  (In addition, if you're running from the build directory, you might have to strip off a ".libs/" added to argv[0] as an artifact of the libtoolizing script.)
Use that in the About dialog, and use it to construct the path of
dumpcap.

Don't put quotes into the path of dumpcap; you don't have to quote
strings with spaces in them when handing them to execvp and, in fact,
you *mustn't* quote them, as the quotes will be treated as part of the
pathname.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17267 f5534014-38df-0310-8fa8-9805f1628bb7