Pass two strings in capture child messages, so the child can send primary and secondary error messages and let the parent worry about how to display them.  This means dumpcap doesn't need stub routines for generating the formatting tags for the primary and secondary messages.
Have a separate message for capture filter errors, so that the parent
can check whether the capture filter looks like a display filter and
report the appropriate message.  This means that dumpcap doesn't need a
stub routine for compiling display filters (a stub routine also means
that Ethereal won't do the check for capture filters that look like
display filters!).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17465 f5534014-38df-0310-8fa8-9805f1628bb7