SCSI is fairly unique among the ethereal protocols in that it is sometimes normal for a PDU we receive back froma  device to be truncated. (report luns with allocation length 8 for example)
Therefore it is a bit wrong to mark these packets as [malformed packets]
Since they are truncated by scsi   and this is NOTY an error condition.


Add a new exception type  :  ScsiBoundsError
If this exception is caught by packet-frame, then print an appropriate message
instead of [malformed packet]

For SCSI, add helper macros TRY_SCSI_SHORT_PACKET   and END_...


If the packet was not short in the normal sense (snaplen < packetlen) then intercept the exception for BoundsError and rethrow it as ScsiBoundsError instead.






git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17611 f5534014-38df-0310-8fa8-9805f1628bb7